Parser.cpp: Fix undefined behaviour in float conversion
The code was trying to check whether a double value could be represented
as an integer by casting it and comparing the result before and after.

However, casting a double value that's out-of-range of the integer is
undefined behaviour, and so in this case Clang on some architectures
would effectively optimize away the range check.

To do this properly, the code has to check that the value will fit in
the integer before casting.

Bug: chromium:845816
Change-Id: I57a723b6e40a1366355030590e70fff19a1159eb