[DevTools] Support the binary type in .pdl files.
In .pdl files, we may then specify 'binary' as type instead
of 'string', indicating that for JSON, it's a base64 encoded
payload that's being transported as a string.

For C++ bindings, the protocol layer will then take care
of the encoding / decoding, so therefore, code_generator.py
references a new protocol::Binary type, which holds a
vector of uint8 internally, with variations to avoid
copying (e.g. reference counted buffer, etc.).
I'm adding this type to the string support libraries
that are part of the inspector protocol - e.g.
v8_inspector_string.h in Chromium.

The important part is that the binary type
- can be instantiated from a protocol::String (by base64 decoding),
- can be converted to a protocol::String (by base64 encoding), and
- provides access to the binary payload via ::data() and ::size().

convert_protocol_to_json.py now includes a backwards compatibility
escape hatch --map_binary_to_string, which makes it so that
in the .json protocol description files, all fields of
type 'binary' remain 'string' for now. I will enable this mode
for the protocol viewer, by adding the flag here:
https://github.com/ChromeDevTools/devtools-protocol/blob/master/scripts/update-to-latest.sh#L38-L41%5C

For Chromium, the .json protocol description files will contain the
binary type because these files are generated and consumed by
code generators, so in the Chromium / v8 code base we won't set the
compatibility flag.

My overall WIP change is in here, and happy to explain more:
https://chromium-review.googlesource.com/c/chromium/src/+/1244719

Bug: chromium:891377
Change-Id: I386014d22a3882541282e7d431f42f05c12643a5