Patch a leading 0 before the decimal dot in JSON double renderings if needed.
It turns out valid JSON requires a 0 before the decimal dot.
While std::stringstream (which we test with) does this "correctly",
base::NumberToString doesn't. So now we're patching it up.

I understand that in some sense, the StrToD should be doing that for
optimal performance in case it's never needed, but I figure it's better
to make it robust just in case and have the code here.

Change-Id: Iaa4f0596c660f438b232b67414ffb70c342c9b43