When converting from json, store 7 bit strings as STRING (major type 3).
So, in JsonToBinaryEncoder::HandleString, we're now checking whether
all characters are in fact 7 bit characters. If so, this string
will become a STRING on the wire.

Conversely, when reading from CBOR / binary, we allow both STRING
and BYTE_STRING. New error codes:
BINARY_ENCODING_INVALID_STRING8: Violating CBOR STRING format
BINARY_ENCODING_STRING8_MUST_BE_7BIT: Violating the additional
constraint that our STRING instances must be 7 bit only.

BINARY_ENCODING_INVALID_MAP_KEY is now used only
if a map key is something other than a STRING or BYTE_STRING.

And there is one bugfix for DecodeUTF8String, I accidentally made
this go to the end of the entire message instead of the end
of the string. It surfaced as soon as I started using it
for the roundtrip tests.

Change-Id: Ibf8cdbc0201635a0efa2eae0dea0bac62c9c98fc