Introduce a CBORTokenizer abstraction, plus refactorings.
CBORTokenizer distinguishes the tokens in a message that a parser
may care about, that is, the scaffolding (MAP_START, STOP)
as well as INT32, STRING8, etc.

The ParseCBOR routine now uses CBORTokenizer to implement
a parser which sends events to a JsonParserHandler.

The Decode* routines are no longer needed.

I've also cleaned the terminology a bit, so that
we call things the same everywhere within this library.
E.g., INT32, DOUBLE, STRING8, STRING16, etc.

Change-Id: Ife975c2344b7f0bc2955dfa26b1ff144301ea5da