BUILD adjustments for easy integration into Chromium tree.
This makes it so that //testing/gmock, //testing/gtest,
//base, and //testing/test.gni have the same path names
that they have in the Chromium tree. This allows us
to build standalone but also within Chromium, without
having to add configuration variables or similar.
Instead, the top-level BUILD.gn file just works in both
this place and in third_party/inspector_protocol/BUILD.gn.

There's also a small tweak for cbor_test.cc, in Chromium

std::isnan isn't in scope so I'm adding the proper include.
Change-Id: I620c2cf3034df4abe12e45f15913d6e79ba576de