[ubsan] Fix UBSan warnings in Maybe<bool> specialization
The default constructor of MaybeBase<> does not initialize the
m_value field, but the move-constructor reads it, so when moving
a default-constructed instance, the bool-typed field being read
can contain a value that's neither 0 nor 1, which is undefined
behavior. This patch fixes that by always initializing the field.
For consistency, the same fix is applied to the int and double
specializations.

The UBSan warnings can be observed when building V8 with
cflags += [ "-fsanitize=undefined" ] and running its "inspector"
tests.

Bug: v8:3770
Change-Id: I20d164434c6f3ee5cb55dd96058f7293bb5133f3