Add a mechanism for recording an ENVELOPE item.
An ENVELOPE consists of:
* 0xd8 - envelope start byte, a tag with add. info 24
* 0x5a - byte string with 32 bit wide length
* a uint32, encoded in 4 bytes (most significant byte first)
* payload

By default, CBORTokenizer will consider the envelope a single
item, that is, Next() will advance past the entire thing
including the payload. However, EnterEnvelope() lets the client
code look inside and explore.

For now, Both arrays and maps are encoded with an envelope
around them. But we may tweak this later.

Since a message is a map at the top level, that means
it's now wrapped with an envelope, allowing us to extract the
size of the message from the first 6 bytes.
It also gives messages a signature: 0xd8, 0xda.

Change-Id: Ibdfe531c2ea8e3de62ef45da2208a283890208d7