First cut for converting between protocol::Value and CBOR.
This is an adaptation of the ParseCBOR routine in cbor.{h, cc}.
Instead of sending events to the JSONParserHandler, this creates
protocol::Value objects.

In the other direction, the straight-forward encoding choices are made
based on our discussion about the protocol thus far:

- All strings are encoded in utf8;
  calls StringUtil::fromUTF / StringUTF8Adapter for this.
- Envelopes around maps and arrays.

Also adds BinaryValue, which wraps a protocol::Binary.
Also adds protocol::Binary::fromSpan, because the existing creation
methods are specific to blink vs. not blink.

Open issues:
- Error handling; this is similar to json parsing; it just bails
  with nullptr - but there's no place where it's logged / recorded,
  unfortunately.
Change-Id: I80d9fa8784a242f5bb51e1511f554549e93623f1