test: Update googletest and rename TEST_CASE -> TEST_SUITE.
googletest's TEST_CASE series of APIs have been deprecated in favor of
(better named) versions TEST_SUITE. This CL updates the usage in
inspector_protocol.

This CL also updates the standalone googletest to the same commit as
Chromium's, because the current googletest does not have the new
TEST_SUITE APIs.

Bug: chromium:925652
Change-Id: Id940f2bf6e2f72f97fefa650f31336168cc5cda6