Tweaks to the standalone build.
This makes it so that include file rules etc. are observed,
even when we start depending on the top-level BUILD.gn from
within Chromium. I made these tweaks in the Chromium tree,
for this change https://chromium-review.googlesource.com/c/chromium/src/+/1495871
and then completed them here to keep things working.
The main thing that's specific to our standalone thing is
that we need a third_party/googletest directory to mimic
Chromium.

Change-Id: I3bfc3526ff88f3e4257df15e50a3af92fb580c67