Bugfix: When comparing the last byte in the message with the stop byte, avoid comparing a char with a uint8_t. This happens because std::string could be the payload, so then, std::string::back() is a reference to a char, not a uint8_t.
Now, EncodeStop() is a uint8_t with value 255. It is just a constant.

So, when I compare the char interpretation of 255, it's -1, and now
the != yields false.

Fix is to use the span<uint8_t> instead, which gives me uint8_t
for that last byte.

I've fixed the test so that it tests both the std::vector<uint8_t>
and the std::string overloads.

Change-Id: Ia2569f4c1bd0482a83ece1b79af4d585620dcc09