Rework the checks so that they are tight for 32 Bit architectures as well.
This migrates the span implementation to std::size_t, from
std::ptrdiff_t, since alph@ pointed out that the signed comparisons
are harder to get right, and std::span actually uses size_t
as well for its index type at this point (this was different in
2018).

So, the checking is now happening against a constant which is either
size_t (on 32 bit architectures), or size_t >> 2 (on 64 bit
architectures). Thus, when we do length computations we have at least
one extra bit for checking the overflow.

Change-Id: Ibb9c7845b39e5f4a24660be7c9590a3ec8000d37