Support STRING16 in the template when converting CBOR map keys to protocol::Value. This was an oversight, while the JSON<->CBOR converter already supports such strings, we were dropping them in this particular conversion.
Longer term, this code should go away, and unfortunately there is
no test for the template here, but I have verified that it works
for the Puppeteer example given in
https://github.com/GoogleChrome/puppeteer/issues/4443.

I'll look into making a test out of this. It'll be a Chromium or
V8 test though, because those are the places where the template
gets expanded.

Change-Id: I3c998ed774ca4537a89b5493fedf258c6b38a9c6