Add CBORTokenizer::GetEnvelope().
Unlike CBORTokenizer::GetEnvelopeContents(), this span includes the
envelope header. Reason this is useful is that for the bindings,
I'd like to assume for now that messages in the wire format contain
a header, and this means that when constructing a DeferredMessage,
I need to include the envelope header.

This adds a bit of direct test coverage for the CBORTokenizer as well,
which should be a healthy thing.

Change-Id: I1bdada7162b764b741f01e4847af4f516603bb97