Separate the lightweight check for CBOR messages from ParseCBOR.
Thus far, ParseCBOR included a lightweight check that assumes
that what's being parsed is a DevTools message: It ensures
that it's a dictionary.

This doesn't work for all cases where we want to use the parser;
specifically, for v8, we want to be able to convert some arguments
for function calls to JSON, and these arguments aren't complete
DevTools messages.

Therefore, this change puts the lightweight check into a new
function, CheckCBORMessage, and makes it so that ParseCBOR
(and thereby also ConvertCBORToJSON) accepts any value, not
just messages (dicts).

Also in this change is a custom gmock matcher for crdtp::Status,
which gives much better error messages when any of the tests fail.

Rolls:
https://chromium-review.googlesource.com/c/chromium/src/+/2002835
https://chromium-review.googlesource.com/c/v8/v8/+/2002935

Change-Id: I915c064f65e7c7880c7b71a30bb0c65ba7535162