Implement crdtp::ErrorSupport, replacing the jinja2 template.
This implementation avoids materializing strings unless there is
an error. Instead it keeps references to const char, and size_t
values (in case the name of the element is an index). We use
<string> from the standard library to produce the error
message, but present it to the client code as a span<uint8_t>.
This works fine since it's 7 bit US-ASCII only.

The rest of the change contains the adjustments needed in the
templates to use this. In the StringUtil classes, we can now
delete all StringBuilder and String/number conversion methods.
This is true for base_string_adapter.{h,cc} in this change,
but also for Blink and V8 utilities downstream.

Chromium Roll:
https://chromium-review.googlesource.com/c/chromium/src/+/2037841

V8 Roll:
https://chromium-review.googlesource.com/c/v8/v8/+/2037959

Change-Id: I3712041cb6ce586c2d751e1ecfeea73c3f1553c1