Introduce a crdtp/dispatch.{h,cc} library.
This replaces the Dispatcher_base_{h,cpp}.template and the
FrontendChannel_h.template.

Apart from where the code lives, the main difference is that
we no longer parse messages in their entirety to dispatch them.
Instead, we perform a shallow parse of only the top-level fields.
This is implemented in crdtp/dispatch.h's Dispatchable class,
which is instantiated from a CBOR message that is a span of bytes.
If the shallow parse is successful, that is, Dispatchable::ok()
is true, then we may call UberDispatcher::Dispatch with it,
which returns a DispatchResult with a Run method on it.

In addition to this, errors are now handled comprehensively,
this includes JSON<->CBOR conversion errors, shallow parsing
errors, as well as other protocol errors.

In content::DevToolsSession (see downstream PR), we send the
errors to the client in the encoding that the client expects.
In other places, we can assume that the content layer has
already checked the message, so now the LOG(ERROR) can be
a DCHECK.

Method not found errors are handled in this way as well -
the UberDispatcher::DispatchResult will send such an error if
executed; however it's also possible to query DispatchResult
as to whether the method was found - this is needed
for implementing fall through between layers.

All protocol error messages are now in UTF8, so agents
/ handlers / etc. in Blink and V8 are adjusted in the downstream
PRs.

In addition to this, protocol responses are now directly
serialized to cbor all the way - there is no more
indirection via ValueConversions. This reduces the Android
binary size by 35k (when the patch is only applied to
Chromium) and should speed things up as well.

Chromium:
https://chromium-review.googlesource.com/c/chromium/src/+/2047966

V8:
https://chromium-review.googlesource.com/c/v8/v8/+/2076215

Change-Id: I20a1e4fce7ecd522d47710c6ee954a180bebe398