For the shallow parse of a DevTools message, allow "params": null.
The params parameter is optional, and there are de-facto three
ways to express absence of this parameter:
1) No "params" key.
2) "params": null
3) "params": {}
Thus far, crdtp::Dispatchable did not support 2), with this
change it'll support it.

This fixes the problem that required this revert:
https://chromium-review.googlesource.com/c/chromium/src/+/2103517

The alternative is to make it so that
content/browser/devtools/protocol/target_handler.cc's checks
if (params && !params->is_none()) before adding a params entry.
SendProtocolMessageToPage, but I think it's better to
also support 2) above, otherwise we may break devtools clients.

Change-Id: Id8122bab6162b2e9c53d70d942366f5a6661ce99